- Adjust NaN replacement query-fragment to set case-insensitivity flag using the "flags" param of regexp_replace()   - That prevents confusion with the question marks sign to be interpreted as query-parameter placeholder - Add implementation for DownloadUrl#store() - Simplify EMR step handling in GraphTransformer - Add default implementation for LambdaBasedStep#init() so that only when needed it must be overridden by subclasses - Use standard method #registerOutputs() in RunEmrJob to upload the generated files - Fix RunEmrJob#validate() method to not throw a validation exception for missing input files but just report not being ready instead - Undo renaming of Step RunEmrJob - Enhance ExportSpaceToFiles step to use GetFeaturesByGeometryBuilder - Add new method SQLQuery#toExecutableQueryString()
